+++ HISTORY.md
@@ -4,3 +4,4 @@
  - 2022-10-04 - hex-gui: add file hex dump macOS Monterey SwiftUI program
  - 2022-10-06 - MyIP: add display public IP address macOS Monterey SwiftUI program
  - 2022-10-24 - SwiftRequest: displays details on the HTTP request, container environment and directory contents for a Swift container deployed to Google Cloud Run
+ - 2022-11-27 - Gist-List: duplicates the GitHub CLI command: `gh gist list`